+++ src/compas_fab/backends/ros/backend_features/move_it_append_collision_mesh.py
@@ -53,5 +53,5 @@ class MoveItAppendCollisionMesh(AppendCollisionMesh):
         co.operation = CollisionObject.APPEND
         world = PlanningSceneWorld(collision_objects=[co])
         scene = PlanningScene(world=world, is_diff=True)
-        request = dict(scene=scene)
+        request = scene.to_request(self.ros_client.ros_distro)
         self.APPLY_PLANNING_SCENE(self.ros_client, request, callback, errback)