+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Unreleased
 
 * Fixed PyBullet loading of meshes.
 * Fixed missing flag in reset planning scene call.
+* Fixed issue on cartesian and kinematic planning when model contains passive joints.
 
 **Deprecated**