+++ CHANGELOG.rst
@@ -15,7 +15,7 @@ Unreleased
 * Added ``compas_fab.backends.PyBulletClient.load_ur5()`` method to simplify some examples.
 * Added ``compas_fab.robots.ReachabilityMap``
 * Added ``compas_fab.robots.DeviationVectorsGenerator``
-* Added ``compas_fab.robots.reachability_map.OrthonormalVectorsFromAxisGenerator``
+* Added ``compas_fab.robots.OrthonormalVectorsFromAxisGenerator``
 
 **Changed**